requirements.txt
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
audioread==2.1.4
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
certifi==2016.9.26
chardet==2.0.1
command-not-found==0.3
configparser==3.5.0
cycler==0.10.0
Cython==0.25.2
debtagshw==0.1
decorator==4.0.10
defer==1.0.6
dirspec==13.10
django-bootstrap3==8.2.3
django-twitter-bootstrap==3.3.0
entrypoints==0.2.2
enum34==1.1.6
fastdtw==0.3.0
functools32==3.2.3.post2
get==0.0.11
gyp==0.1
httplib2==0.8
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
joblib==0.10.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
librosa==0.4.3
lxml==3.3.3
MarkupSafe==0.23
matplotlib==1.5.3
mercurial==2.8.2
mistune==0.7.3
mpmath==0.19
mysql-connector-python==1.1.6
mysql-utilities==1.3.5
nbconvert==4.2.0
nbformat==4.2.0
nose==1.3.7
notebook==4.2.3
numpy==1.8.2
oauthlib==2.0.0
oneconf==0.3.7.14.4.1
PAM==0.4.2
pandas==0.19.1
paramiko==1.10.1
pathlib2==2.1.0
pexpect==3.1
pickleshare==0.7.4
Pillow==2.3.0
piston-mini-client==0.7.5
post==0.0.8
prompt-toolkit==1.0.9
psutil==1.2.1
ptyprocess==0.5.1
public==0.0.34
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pydub==0.16.6
Pygments==2.1.3
pygobject==3.12.0
pygpgme==0.3
pyOpenSSL==0.13
pyparsing==2.1.10
pyserial==2.6
pysmbc==1.0.14.1
pysqlite==2.6.3
python-apt===0.9.3.5ubuntu2
python-dateutil==2.6.0
python-debian===0.1.21-nmu2ubuntu2
python-speech-features==0.4
pytz==2016.10
pyxdg==0.25
pyzmq==16.0.2
qtconsole==4.2.1
query-string==0.0.8
reportlab==3.0
request==0.0.8
requests==2.11.1
requests-oauthlib==0.6.2
resampy==0.1.4
scikit-learn==0.18.1
scipy==0.18.1
selenium==3.3.3
sessioninstaller==0.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
software-center-aptd-plugins==0.0.0
sympy==1.0
system-service==0.1.6
terminado==0.6
tornado==4.4.2
traitlets==4.3.1
tweepy==3.6.0
Twisted-Core==13.2.0
Twisted-Web==13.2.0
unity-lens-photos==1.0
urllib3==1.7.1
virtualenv==15.0.2
wcwidth==0.1.7
widgetsnbextension==1.2.6
xdiagnose===3.6.3build2
zope.interface==4.0.5
